+++ test/devices/deviceSettings.test.ts
@@ -10,7 +10,6 @@ import {
 } from '../../src';
 import ExampleDevices from './exampleDevices.json';
 
-jest.mock('unifi-protect', () => jest.fn()); // Working now, phew
 jest.mock('unifi-access', () => jest.fn()); // Working now, phew
 
 describe('Device Settings', () => {