+++ test/heater/temperaturSetttings.test.ts
@@ -1,6 +1,5 @@
 import { Daytime, TemperatureSettings } from '../../lib';
 
-jest.mock('unifi-protect', () => jest.fn()); // Working now, phew
 jest.mock('unifi-access', () => jest.fn()); // Working now, phew
 
 describe('TemperatureSettings', () => {