+++ test/services/muell-service.test.ts
@@ -1,6 +1,5 @@
 import { MuellService, ServerLogService } from '../../src';
 
-jest.mock('unifi-protect', () => jest.fn()); // Working now, phew
 jest.mock('unifi-access', () => jest.fn()); // Working now, phew
 
 describe('MuellService', () => {