+++ test/services/time-callback-service.test.ts
@@ -8,7 +8,6 @@ import {
   TimeOfDay,
 } from '../../src';
 
-jest.mock('unifi-protect', () => jest.fn()); // Working now, phew
 jest.mock('unifi-access', () => jest.fn()); // Working now, phew
 
 describe('TimeCallbackService', () => {